With this feature branch we close #13 and solve #14

+++ src/test/java/es/tid/keyserver/https/certificate/HttpsCertTest.java
@@ -68,7 +68,7 @@ public class HttpsCertTest {
         System.out.println("certRemainDays");
         long result = instance.certRemainDays();
         System.out.println("[ TEST ] Certificate validity remain days: " + result);
-        assertNotNull((Long)result);
+        assertNotNull(result);
     }
 
     /**
